Peer添加函数&字段类型修改为uint

+++ CSharpENet/CSharpENet/Channel.cs
@@ -2,12 +2,12 @@
 
 struct ENetChannel
 {
-    public ushort outReliableSeqNumber = 0;
-    public ushort outUnreliableSeqNumber = 0;
-    public ushort usedReliableWindows = 0;
-    public ushort[] rSeliableWindows = new ushort[(int)ENetDef.PeerReliableWindows];
-    public ushort inReliableSeqNumber = 0;
-    public ushort inUnreliableSeqNumber = 0;
+    public uint outReliableSeqNumber = 0;
+    public uint outUnreliableSeqNumber = 0;
+    public uint usedReliableWindows = 0;
+    public uint[] rSeliableWindows = new uint[(int)ENetDef.PeerReliableWindows];
+    public uint inReliableSeqNumber = 0;
+    public uint inUnreliableSeqNumber = 0;
     public LinkedList<ENetInCmd> inReliableCmds = new();
     public LinkedList<ENetInCmd> inUnreliableCmds = new();
